Tweaks for OAE-isms (no handling floats in content props, cannot return null from handle when no changes are desired)

+++ src/main/java/net/unicon/kaltura/MediaItem.java
@@ -180,8 +180,8 @@ public class MediaItem implements Serializable {
         return MediaItem.truncateText(getDesc(), shortMaxDesc, shortSpace);
     }
 
-    public float getDuration() {
-        float duration = 0;
+    public int getDuration() {
+        int duration = 0;
         if (kalturaItem != null) {
             if (kalturaItem instanceof KalturaMediaEntry) {
                 duration = ((KalturaMediaEntry) kalturaItem).duration;
